% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G.matrix.R
\name{G.matrix}
\alias{G.matrix}
\title{Generate Genetic Relationship Matrix}
\usage{
G.matrix(M, method=c("VanRaden", "UAR", "UARadj"), format=c("wide", "long"))
}
\arguments{
\item{M}{\code{matrix}. Matrix of markers in which \eqn{n} individuals are in rows and \eqn{p} markers in columns.}

\item{method}{\code{character}. Method for constructing the GRM. Three methods are currently supported. \code{"VanRaden"} indicates the method proposed by Vanraden (2008) for additive
and dominant genomic relationship. \code{"UAR"} and \code{"UARadj"} represent methods proposed by Yang et al. (2010) for additive genomic relationship. See 'Detais'}

\item{format}{\code{character}. Class of object to be returned. \code{"wide"} returns a \eqn{n} x \eqn{n} matrix.
\code{"long"} returns the GRM as a table with 3 columns. See 'Details'}
}
\value{
The GRM is returned in a matrix or table format.
}
\description{
G.matrix allows to create Genomic Relationship Matrix (GRM)
}
\details{
Function G.matrix provides three diferent types of relationship matrix. For \code{"VanRaden"} method, the relationship matrix is estimated as proposed by Vanraden (2008):
 \deqn{G = \frac{MM'}{2\sum p_i(1-p_i)}} where: \eqn{M} is the marker matrix. The SNP genotype takes a value of 0, 1 or 2 if the genotype of the \eqn{j}th
individual at SNP \eqn{i} is \eqn{aa}, \eqn{Aa} or \eqn{AA}, respectively.
\code{"UAR"} and \code{"UARadj"} represents the genomic relationship matrix proposed by Yang et al. (2010) named by Powell et al. (2010) as Unified Additive Relationship (UAR)
and Adjusted UAR, respectively. GRM is then obtained by:
\deqn{A_{jk} = \frac{1}{N}\sum_i{A_{ijk}} = \begin{cases}
\dfrac{1}{N} \sum_i{\dfrac{(x_{ij} - 2p_{i})(x_{ik} - 2p_i)}{2p_i(1-p_i)}}, j \neq k \cr
1 +  \dfrac{1}{N} \sum_i{\dfrac{x_{ij}^{2}(1 + 2p_{i})x_{ij} + 2p_i^{2}}{2p_i(1-p_i)}}, j = k
\end{cases}}
where: \eqn{p_i} is the allele frequency at SNP \eqn{i}, \eqn{x_{ij}} is the SNP genotype that takes a value of 0, 1 or 2 if the genotype of the \eqn{j}th
individual at SNP \eqn{i} is \eqn{aa}, \eqn{Aa} or \eqn{AA}, respectively.
The \code{format} argument is the desired output format. For \code{"wide"}, the relationship output produced is in matrix format, with \eqn{n x n} dimension. 
If \code{"long"} is the chosen format, the inverse of the relationship matrix is produced and converted to a table. In this case, the upper triangular part of the relationship matrix
is changed to a table with 3 columns representing the respective rows, columns and values (Used mainly by ASReml).
If the relationship matrix is not positive definite, a near positive definite matrix is created and solved, followed by a warning message.
}
\examples{

#(1) Additive and dominance relationship matrix 
data(maize.hyb)
x <- G.matrix(maize.hyb, method="VanRaden", format = "wide")
A <- x$Ga
D <- x$Gd

}

