% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize.hyb}
\alias{maize.hyb}
\title{maize hybrids}
\format{A matrix with 50 rows and 492 columns, where hybrids are in rows and SNP markers in columns}
\description{
50 hybrids of maize genotyped with 492 SNP markers
}
\examples{
#' data(maize.hyb)


}
\keyword{datasets;}
\keyword{hybrids}
