% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize.line}
\alias{maize.line}
\title{maize lines}
\format{A matrix with 70656 observations on the following 4 variables.
\itemize{ 
\item{sample: }{identification of samples (name of individuals)}
\item{marker: }{identification of SNP markers}
\item{allele.1: }{Allele 1}
\item{allele.2: }{Allele 2}}}
\source{
Lines genotyped from allogamous breeding laboratory - ESALQ/USP \url{http://www.genetica.esalq.usp.br/alogamas/index2.html}
}
\description{
A raw dataset of maize lines genotyped with 768 markers
}
\examples{

data(maize.line)
## str(maize.line)

}
\keyword{datasets;}
\keyword{lines}
