\name{snpReady-package}
\alias{snpReady-package}
\alias{snpReady}
\docType{package}
\title{
\packageTitle{snpReady}
}
\description{
\packageDescription{snpReady}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{snpReady}
\packageIndices{snpReady}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{snpReady}

Maintainer: \packageMaintainer{snpReady}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
