% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen.R
\name{popgen}
\alias{popgen}
\title{Estimate parameters of population genetics from genomic data}
\usage{
popgen(Z, subgroups)
}
\arguments{
\item{Z}{object of class \code{matrix} a (non-empty) matrix of markers incidence
by genotype. The columns represent the markers and rows the individuals}

\item{subgroups}{\code{vector}. Vector with information of subgroups or
subpopulations for individuals.}
}
\value{
The function returns two lists, one with general information for markers and individuals and another by group (if this is available).
\code{General}
for each marker: the function returns allelic frequency (as \code{p} and \code{q}),
Minor Allele Frequency (MAF), expected heterozygosity (He), observed
heterozygosity (Ho), Nei's Genetic Diversity (DG) and Polymorphism Informative Content(PIC). 
For population:  Same parameters used for each marker are returned for general population with mean, lower and upper limits.
For individuals: it's returned his observed heterozygosity (Ho), coefficient of inbreeding (Fi) and selfing index (Si).

Variability: shows estimates of effective population size (Ne), variance adittive component(Va) and variance dominance component (Vd) and a
summary of number of groups, genotypes and markers.

\code{bygroups} 
Same output produced in general list, here is generated for subpopulations or subgroups. Moreover, for subgroups is outputed
the number of exclusive alleles and number of fixed alleles per group
}
\description{
This function allows estimate parameters of population genetics from genomic data. Further,
it also allows estimate those parameters by subpopulations.
}
\details{
The matrix of makers are the dimension nxm, where individuals are in rows and markers in column. The
genotypic matrix should be numeric and it should be contain only the name of
the markers in the column.
The number of subgroups is defined by the user and it can accept anything to distinguish subpopulations.
}
\examples{
# hybrids maize data
M <- data(maize.hyb)
x <- popgen(M) 

# using subpopulations
PS<-c(rep(1,25), rep(2,25))
x <- popgen(M, subgroups=PS)
}

