% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen.R
\name{popgen}
\alias{popgen}
\title{Population genetics from genomic data}
\usage{
popgen(Z, subgroups)
}
\arguments{
\item{\code{Z}}{object of class \code{matrix}. A (non-empty) matrix of molecular markers. Markers must be in columns and individuals in rows.}

\item{\code{subgroups}}{a \code{vector}. Vector with information for subgroups or subpopulations.}
}
\value{
Two lists are returned, one with general information for markers and individuals and another by group (if applicable).

\code{General}

For each marker: allelic frequency (\code{p} and \code{q}),
Minor Allele Frequency (\eqn{MAF}), expected heterozygosity (\code{He}), observed
heterozygosity (\eqn{Ho}), Nei's Genetic Diversity (\eqn{DG}) and Polymorphism Informative Content(\eqn{PIC}). 

For genotypes:  observed heterozygosity (\eqn{Ho}), coefficient of inbreeding (\eqn{Fi}) and selfing index (\eqn{Si}) are returned.

For population:  parameters used for markers are returned for general population with mean, lower and upper limits.

Variability: shows estimates of effective population size (\eqn{Ne}), adittive (\eqn{Va}) and dominance \eqn{Vd} variance components, and a
summary of number of groups, genotypes and markers.

\code{bygroups}

Same outputs are here generated for subpopulations or subgroups. Moreover, number of exclusive and fixed alleles per group for subgroups are also returned.
}
\description{
This function allows for estimating parameters of population genetics from genomic data. In addition,
it also allows estimations considering subpopulations.
}
\details{
The matrix of makers is of dimension \eqn{n x p}, in which individuals are in rows and markers in columns.
The number of subgroups is defined by the user and accepts any data type (\code{character}, \code{integer}, \code{numeric}...) to distinguish subpopulations.
}
\examples{
# hybrid maize data
M <- data(maize.hyb)
x <- popgen(M) 

# using subpopulations
PS<-c(rep(1,25), rep(2,25))
x <- popgen(M, subgroups=PS)
}

