% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.data.R
\name{raw.data}
\alias{raw.data}
\title{Preparation of genomic data}
\usage{
raw.data(data, frame=c("long","wide"), hapmap,
       bases=TRUE, sweep.sample= 1, call.rate=0.95, maf=0.05,
       input=TRUE, outfile=c("012","-101","structure"))
}
\arguments{
\item{data}{object of class \code{matrix}}

\item{frame}{\code{character}. Format of genomic data to be inputed. Two formats are currently supported. \code{"long"} is used for objects
with sample ID (1st column), marker ID (2nd column), fist allele (3rd column) and second allele (4th column). \code{"wide"} inputes
a \eqn{n} x \eqn{m} matrix where markers must be in columns and individuals in rows}

\item{hapmap}{\code{matrix}. Object with information on SNPs, chromosome and position}

\item{base}{\code{logical}. Are gentype coded as nitrogenous bases? if \code{TRUE}, data are converted to numeric. If \code{FALSE}, it follows to clean up.}

\item{sweep.sample}{\code{numeric}. Threshold for removing samples from data by missing rate. Samples with missing rate above the defined threshold are
removed from dataset.}

\item{call.rate}{\code{numeric}. Threshold for removing marker by missing genotype rate. SNP with \code{"call rate"} below threshold are removed from dataset.}

\item{maf}{\code{numeric}. Threshold for removing SNP by minor allele frequency.}

\item{input}{\code{logical}. If \code{"TRUE"}, imputation of missing data is performed. See details.}

\item{outfile}{\code{character}. Type of output to be produced. \code{"012"} outputs matrix coded as 0 to \code{AA}, 1 to \code{Aa} and 2 to \code{aa}. \code{"-101"}
presents marker matrix coded as -1, 0 and 1 to \code{aa}, \code{Aa} and \code{AA}, respectively. \code{"structure"} returns a matrix suitable for STRUCTURE Software.
For this, each remaining marker is splited in two columns, one for each allele. Nitrogenous bases are then recoded to a number, so A is 1, C is 2, G is 3 and T is 4.}
}
\value{
Returns a properly coded marker matrix output and a report specifying which individuals are removed by \code{sweep.sample} and which markers are removed by \code{"call.rate"}
and \code{"maf"}.
}
\description{
This function gets genomic data ready to be used in packages or softwares that
perform genomic predictions
}
\details{
The function allows flexible imputation of genomic data. Data might be in long format with 4 columns or in wide 
format where markers are in columns and individuals in rows. Both numeric and nitrogenous bases are accepted. Samples and markers can be eliminated based on missing data rate. Markers can also be eliminated based on
the frequency of the minor allele. Imputation is carried out through combination of allelic frequency and individual
inbreeding coefficient. Hence, for missing values, genotypes are imputed based on their probability of occurrence. This probability
depends both on genotype frequency and inbreeding of the individual a specific locus.
}
\examples{
data(maize.line)
raw.data(maize.line, frame="table", base=TRUE, hapmap, sweep.sample= 1, 
         call.rate=0.95, maf=0.05, input=TRUE, outfile="-101")


}
\references{
# missing
}
\seealso{
# # missing
}

