\name{SnpReady-package}
\alias{SnpReady-package}
\alias{SnpReady}
\docType{package}
\title{
\packageTitle{SnpReady}
}
\description{
\packageDescription{SnpReady}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SnpReady}
\packageIndices{SnpReady}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SnpReady}

Maintainer: \packageMaintainer{SnpReady}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
